feat: implement import workflow components and refactor main execution logic
