+++ scripts/send/send-import-notifications/src/main.ts
@@ -6,43 +6,11 @@
  */
 
 import { Core } from '@go-automation/go-common';
-import {
-  SENDNotifications,
-  SENDNotificationImportWorker,
-  QATestFormatAdapter,
-  loadUploadedAttachments,
-} from '@go-automation/go-send';
-import type { SENDNotificationRow, SENDUploadedAttachment } from '@go-automation/go-send';
 
 import { setupEventListeners } from './libs/setupEventListeners.js';
-import { displayResults } from './libs/displayResults.js';
 import { handleWorkflowError } from './libs/handleWorkflowError.js';
+import { createImportWorkflowComponents, executeImportWorkflow } from './libs/importWorkflow.js';
 import type { ImportNotificationsConfig } from './types/ImportNotificationsConfig.js';
-
-/** Base export columns for the CSV output */
-const BASE_EXPORT_COLUMNS = [
-  'ID_Scenario',
-  'Scenario',
-  'Prodotto',
-  'Destinatario',
-  'Denomination',
-  'Indirizzo PEC',
-  'physicalCommunicationType',
-  'CAP',
-  'Provincia',
-  'Citta',
-  'Stato',
-  'Range',
-  'Indirizzo',
-  'Sender',
-  'Tax ID',
-  'RequestID',
-  'Data invio Test',
-  'Stato',
-  'Esito',
-  'Note',
-  'iun',
-];
 
 /**
  * Main script execution function.
@@ -54,116 +22,13 @@ const BASE_EXPORT_COLUMNS = [
  */
 export async function main(script: Core.GOScript): Promise<void> {
   const config = await script.getConfiguration<ImportNotificationsConfig>();
+  const components = await createImportWorkflowComponents(script, config);
 
-  // Resolve export file path
-  const exportPathInfo = script.paths.resolvePathWithInfo(config.exportFile, Core.GOPathType.OUTPUT);
-  if (exportPathInfo) {
-    if (exportPathInfo.isAbsolute) {
-      script.logger.info(`Export path (absolute): ${exportPathInfo.path}`);
-    } else {
-      script.logger.info(`Output directory: ${exportPathInfo.resolvedDir}`);
-      script.logger.info(`Export file will be saved to: ${exportPathInfo.path}`);
-    }
-  }
-
-  // Initialize SDK
-  script.logger.section('Initializing SEND SDK');
-  const sdk = new SENDNotifications({
-    basePath: config.basePath,
-    apiKey: config.pnApiKey,
-    timeout: 300000,
-    debug: false,
-    ...(config.proxyUrl !== undefined && { proxyUrl: config.proxyUrl }),
-  });
-
-  if (config.proxyUrl) {
-    script.logger.info(`Using HTTP proxy: ${config.proxyUrl}`);
-  }
-  script.logger.success('SDK initialized');
-
-  // Load pre-uploaded attachments (send-upload-attachments results), grouped by pratica
-  let attachmentsByPratica: ReadonlyMap<string, ReadonlyArray<SENDUploadedAttachment>> | undefined;
-  if (config.attachmentsFile) {
-    script.logger.section('Loading Uploaded Attachments');
-    const attachmentsPath = script.paths.resolvePath(config.attachmentsFile, Core.GOPathType.INPUT);
-    script.logger.info(`Attachments file: ${attachmentsPath}`);
-
-    const loadResult = await loadUploadedAttachments(attachmentsPath);
-    attachmentsByPratica = loadResult.attachmentsByPratica;
-
-    for (const skipped of loadResult.skipped) {
-      script.logger.warning(
-        `Skipped attachment "${skipped.filePath || 'unknown'}" (pratica: ${skipped.pratica || 'n/a'}): ${skipped.reason}`,
-      );
-    }
-    if (loadResult.totalAttachments === 0) {
-      throw new Error(`No usable attachments found in ${attachmentsPath}`);
-    }
-    script.logger.success(
-      `Loaded ${loadResult.totalAttachments} attachments for ${attachmentsByPratica.size} pratiche`,
-    );
-  }
-
-  // Create importer
-  script.logger.section('Setting up CSV Importer');
-  const importerAdapter = new QATestFormatAdapter();
-  const importer = new Core.GOCSVListImporter<SENDNotificationRow>({
-    ...importerAdapter.getOptions(),
-    preserveOriginalData: config.preserveAllColumns,
-  });
-
-  // Create exporter (if export file specified)
-  let exporter: Core.GOCSVListExporter<Record<string, unknown>> | undefined;
-  if (exportPathInfo) {
-    script.logger.info(`Export file: ${exportPathInfo.path}`);
-    exporter = new Core.GOCSVListExporter({
-      outputPath: exportPathInfo.path,
-      includeHeader: true,
-      delimiter: ',',
-      columns: BASE_EXPORT_COLUMNS,
-      mergeOriginalColumns: config.preserveAllColumns,
-      columnConflictStrategy: 'keep-generated',
-      skipInvalidItems: true,
-    });
-
-    if (config.preserveAllColumns) {
-      script.logger.info('CSV passthrough enabled: all original columns will be preserved');
-    }
-  }
-
-  // Create worker and register events
-  const worker = new SENDNotificationImportWorker(importer, sdk);
-  setupEventListeners(worker, importer, exporter, script.prompt);
+  setupEventListeners(components.worker, components.importer, components.exporter, script.prompt);
   script.logger.success('Components initialized');
 
-  // Execute workflow
-  script.logger.section('Starting Import Workflow');
-  const csvInputPath = script.paths.resolvePath(config.csvFile, Core.GOPathType.INPUT);
-  script.logger.info(`Input file: ${csvInputPath}`);
-  script.logger.info(`Send mode: ${config.sendNotifications ? 'LIVE' : 'DRY-RUN'}`);
-  script.logger.info(`Concurrency: ${config.concurrency}`);
-  script.logger.info(`Poll for IUN: ${config.pollForIun}`);
-  script.logger.newline();
-
   try {
-    const result = await worker.process(csvInputPath, {
-      concurrency: config.concurrency,
-      skipFailedNotifications: false,
-      sendNotifications: config.sendNotifications,
-      pollForIun: config.pollForIun,
-      pollMaxAttempts: config.pollMaxAttempts,
-      pollDelayMs: config.pollDelayMs,
-      useStreaming: 'auto',
-      streamingThresholdMB: config.streamingThresholdMb,
-      exporter: exporter,
-      preserveAllColumns: config.preserveAllColumns,
-      exportAllRows: config.exportAllRows,
-      includeStatusColumns: config.includeStatusColumns,
-      ...(attachmentsByPratica !== undefined && { attachmentsByPratica }),
-    });
-
-    displayResults(script, result, exportPathInfo?.path);
-    script.logger.success('Workflow completed successfully');
+    await executeImportWorkflow(script, config, components);
   } catch (error) {
     handleWorkflowError(error, script.logger);
     throw error;
